% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_add_dem_height}
\alias{cqcp_add_dem_height}
\title{Add DEM height to data.table.}
\usage{
cqcp_add_dem_height(
  data,
  file = NULL,
  raster = NULL,
  directory = NULL,
  outfile = NULL,
  overwrite = TRUE,
  crop = FALSE,
  na_vals = 0,
  ...
)
}
\arguments{
\item{data}{data.table/frame with at least columns 'lon' and 'lat'}

\item{file}{Path to a DEM geotiff}

\item{raster}{RasterLayer object with DEM data}

\item{directory}{Directory path to store the downloaded SRTM data. If NULL,
downloaded data is stored in the current working directory.}

\item{outfile}{File path to save the created SRTM raster as geotiff}

\item{overwrite}{overwrite existing geotiff?}

\item{crop}{Crop SRTM raster/geotiff to data extent}

\item{na_vals}{Set NA values in DEM to this value to avoid missing value in m2. Default: 0}

\item{...}{Additional parameters supported by raster::getData}
}
\value{
data table with new column 'z' with DEM information
}
\description{
Adds Digital Elevation Model (DEM) information from a RasterLayer Object or 
geotiff to the data.table.
If no RasterLayer Object or file are given as parameters, SRTM data is 
automatically downloaded and used to extract DEM information at the positions
of the stations.
A new column 'z' is added to the data.table to be used in filter level M2.
}
