% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_add_episode}
\alias{cqcp_add_episode}
\title{add_episode}
\usage{
cqcp_add_episode(data, duration)
}
\arguments{
\item{data}{data.table with at least columns 'p_id' and 'time'}

\item{column}{}
}
\value{
data.table with additional column 'episode'
}
\description{
add_episode
}
