% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_colourise}
\alias{cqcp_colourise}
\title{cqcp_colourise}
\usage{
cqcp_colourise(str, colour)
}
\arguments{
\item{str}{A string.}

\item{colour}{"red", "green", or "yellow"
This can be, e.g., '10 days'.}
}
\value{
coloured string
}
\description{
Colours a string with ANSI escapes.
}
