% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_cor_month}
\alias{cqcp_cor_month}
\title{Monthly correlation of individual CWS with median}
\usage{
cqcp_cor_month(x, y, m, cutOff)
}
\arguments{
\item{x}{values of unaggregated time series}

\item{y}{data.table containing column 'med' holding aggregated values and
month holding the month the time series belongs to}

\item{m}{month to base the calculation on}

\item{cutOff}{value below which FALSE is returned.}
}
\value{
TRUE if correlation for the given month is higher than cutOff, FALSE
  otherwise
}
\description{
Calculates the correlation of each CWS vs. a data.set containing an
aggregated time series in the column 'med' per month. Both series need to
have the same length and values at the same position are expected to belong
to the same position in time. Function is called internally by QC m4.
}
