% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_cor_timespan}
\alias{cqcp_cor_timespan}
\title{Correlation of individual CWS with median for a certain timespan.}
\usage{
cqcp_cor_timespan(x, y, t, cutOff, timespan = "month")
}
\arguments{
\item{x}{values of unaggregated time series}

\item{y}{data.table containing column 'med' holding aggregated values and
column specified via 'timespan', holding the group the time series belongs to}

\item{t}{timespan (month, or episode number) to base the calculation on}

\item{cutOff}{value below which FALSE is returned.}

\item{timespan}{column name to base the calculations on. Default: 'month'}
}
\value{
TRUE if correlation for the given month is higher than cutOff, FALSE
  otherwise
}
\description{
Calculates the correlation of each CWS vs. a data.set containing an
aggregated time series in the column 'med' per timespan. Both series need to
have the same length and values at the same position are expected to belong
to the same position in time. Function is called internally by QC m4.
This function was formerly cor_month but is now compatible with 'duration' parameter.
}
