% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_download_srtm}
\alias{cqcp_download_srtm}
\title{Download SRTM data using raster package for bounding box of data.}
\usage{
cqcp_download_srtm(
  data,
  directory = NULL,
  outfile = NULL,
  overwrite = TRUE,
  crop = FALSE,
  ...
)
}
\arguments{
\item{data}{data.table/data.frame with at least columns 'lon' and 'lat'}

\item{directory}{directory path to optionally store the SRTM data}

\item{outfile}{file path to save the SRTM raster as geotiff}

\item{overwrite}{overwrite existing geotiff? Default is TRUE.}

\item{crop}{crop raster/geotiff to data extent? Default is FALSE.}

\item{...}{additional parameters supported by raster::getData}
}
\value{
RasterLayer object with SRTM data
}
\description{
SRTM data is automatically downloaded and, in case of more than one SRTM tile, 
merged together as a mosaic.
The SRTM RasterLayer Object can be cropped to the geographical extent of 
the data (crop = TRUE).
SRTM source: https://srtm.csi.cgiar.org/
}
