% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_has_column}
\alias{cqcp_has_column}
\title{has_column}
\usage{
cqcp_has_column(data, column = "month")
}
\arguments{
\item{data}{data.table}

\item{column}{The column to look for. Default: "month"}
}
\value{
true if data contains a specific column
}
\description{
has_column
}
