% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m2}
\alias{cqcp_m2}
\title{Main QC step m2}
\usage{
cqcp_m2(
  data,
  low = 0.01,
  high = 0.95,
  heightCorrection = T,
  debug = F,
  lapse_rate = 0.0065,
  t_distribution = F
)
}
\arguments{
\item{data}{data.table object obtained from m1}

\item{low}{0 < low < high < 1}

\item{high}{0 < low < high < 1}

\item{heightCorrection}{if set to true (default) and the column "z" exists in
the input data, the temperatures used in calculating the z-score are
corrected. The applied formula is ta_cor = ta + ((lapse_rate * (z - mz)) 
where mz is the spatial mean of z at each time step.}

\item{debug}{set to true to keep intermediate results}

\item{lapse_rate}{Lapse rate to use in 'heightCorrection'. Default is the 
environmental lapse rate of -0.0065 K/m. Set as a positive value: e.g.
lapse_rate = 0.01 to set a dry adiabatic lapse rate.}

\item{t_distribution}{Set to TRUE to assume a Student-t distribution of the 
data instead of the normal distribution. Default: FALSE}
}
\value{
data.table
}
\description{
Flags all values as FALSE if robust z-score is not within the critical values
obtained from low and high. This approach is based on the distribution
function of all values. !Careful: if number of stations is < 200, it would be
better to use the Student-t distribution (t_distribution = TRUE)!
}
\examples{
data(netatmoBer)
x <- cqcp_m1(netatmoBer)
y <- cqcp_m2(x)

}
