% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m3}
\alias{cqcp_m3}
\title{Main QC step m3}
\usage{
cqcp_m3(data, cutOff = 0.2, complete = FALSE, duration = NULL, rolling = FALSE)
}
\arguments{
\item{data}{data.table object obtained from m2}

\item{cutOff}{value above which data are flagged with FALSE, 0 < cutOff < 1.
Default is 0.2, i.e., 20 percent of data.}
}
\value{
data.table
}
\description{
Flag values with FALSE if more than cutOff percent values are removed during
m2 per month. This is done since it is assumed that if too many individual
values are flagged FALSE in m2, the station is too suspicious to be kept.
}
\examples{
y <- m2(x)
z <- m3(y)
}
