% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m4}
\alias{cqcp_m4}
\title{Main QC step m4}
\usage{
cqcp_m4(data, cutOff = 0.9, complete = FALSE, duration = NULL, rolling = FALSE)
}
\arguments{
\item{data}{data.table as returned by m3}

\item{cutOff}{value of correlation coefficient below which data are flagged
with FALSE, 0 < cutOff < 1. Default is 0.9.}

\item{complete}{logical. Use the complete data set for the filter (priority over 'duration')?}

\item{duration}{A fixed duration to be used (cf. lubridate duration documentation).
This can be, e.g., '10 days'.}

\item{rolling}{logical. Set to TRUE to carry out the filter on a rolling basis. 
A 'duration' has to be specified too. Not yet implemented.}
}
\value{
data.table
}
\description{
Flag values with FALSE if the correlation with the median of all stations is 
lower than cutOff.
Default is to apply m4 per month, but can be changed to use the complete data
set (e.g. for short data sets), some other fixed duration (e.g. '10 days'), or
rolling, in combination with 'duration' (not yet implemented.)
}
\examples{
data(netatmoBer)
m_1 <- cqcp_m1(netatmoBer)
m_2 <- cqcp_m2(m_1)
m_3 <- cqcp_m3(m_2)
m_4 <- cqcp_m4(m_3)
}
