% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m5}
\alias{cqcp_m5}
\title{Main QC step m5}
\usage{
cqcp_m5(
  data,
  radius = 3000,
  n_buddies = 5,
  alpha = 0.1,
  heightCorrection = TRUE,
  lapse_rate = 0.0065,
  check_elevation = TRUE,
  max_elev_diff = 100,
  keep_isolated = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data}{data.table object as returned by cqcp_m4}

\item{radius}{A radius in meter around each station to check for neighbours.
Default: 3000}

\item{n_buddies}{Minimum number of neighbouring stations with a valid 
value within 'radius' Default: 5}

\item{alpha}{Significance level (two-tailed approach) for the outlier detection 
within 'radius'. 0 < alpha < 1. Default: 0.1.}

\item{heightCorrection}{If set to TRUE (default) and the column "z" exists in
the input data, the temperatures used in calculating the z-score are
corrected. The applied formula is ta_cor = ta + ((lapse_rate * (z - mz)) 
where mz is the spatial mean of z at each time step.}

\item{lapse_rate}{Lapse rate to use in 'heightCorrection'. Default is the 
environmental lapse rate of -0.0065 K/m. Set as a positive value: e.g.
lapse_rate = 0.01 to set a dry adiabatic lapse rate. For consistency, this
lapse rate should be the same as in ?cqcp_m2.}

\item{check_elevation}{Set to TRUE to check whether the elevation in column "z" 
of the neighbouring stations within 'radius' should be compared to the 
elevation of the station that is checked. If the absolute elevation difference 
between each buddy and the station itself is larger than 'max_elev_diff', this 
neighbour is left out of the calculations. Default: TRUE}

\item{max_elev_diff}{Maximum allowed elevation difference in meters between 
the station that is checked and each neighbour within 'radius'. Default: 100 m.}

\item{keep_isolated}{Set to TRUE if isolated stations (with less buddies than 
'n_buddies') should be kept (m5 = TRUE) or removed (m5 = FALSE). Default: FALSE.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Spatial buddy check. Flags all values as FALSE if values are too different
than those from spatially neighbouring stations/buddies (spatial outlier filter).
Check is based on the median value and the Qn estimator, as in cqcp_m2, of 
the buddies within a specified radius. The station that is checked is not
considered when calculating these values.
A minimum number of buddies within the radius has to be present for the 
calculations, thus this filter also flags isolated stations. Set 
'keep_isolated' = TRUE if you want to keep the isolated stations.
In addition to flag variable 'm5' another variable 'isolated' is added to the
data.table with FALSE and TRUE, if stations have >='n_buddies' within 'radius' 
or not, respectively.
}
\examples{
data(cqcp_cws_data)
data <- cqcp_padding(cqcp_cws_data)
data_qc <- cqcp_m1(data)
data_qc <- cqcp_m2(data_qc)
data_qc <- cqcp_m3(data_qc)
data_qc <- cqcp_m4(data_qc)
data_qc <- cqcp_m5(data_qc)
data_qc_keep <- cqcp_m5(data_qc, keep_isolated = TRUE) # keep isolated CWS
}
