% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o1}
\alias{cqcp_o1}
\title{Optional QC step o1}
\usage{
cqcp_o1(data, fun = cqcp_interpol, quiet = FALSE, ...)
}
\arguments{
\item{data}{data.table object as returned by cqcp_m5}

\item{fun}{Function to use for interpolation, default is interpol}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}

\item{...}{Additional parameters for interpolation function}
}
\value{
data.table
}
\description{
In this step missing data is interpolated, default is to perform linear
interpolation on gaps of maximal length = 1.
A new column named "ta_int" is added to the data.table, containing the
filtered data at QC level m5 with the interpolated data.
}
