% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o2}
\alias{cqcp_o2}
\title{Optional QC step o2}
\usage{
cqcp_o2(data, cutOff = 0.8, quiet = FALSE)
}
\arguments{
\item{data}{data.table object as returned by cqcp_o1}

\item{cutOff}{Percentage of values that must be present for each day before
all values of that day are flagged with FALSE, expressed in fraction: 0 <
cutOff < 1. Default is 0.8, i.e, 80 percent of data.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Optional QC for temporal data availability. Flags all values in a
calendar day as FALSE if less than 'cutOff' percent of valid values are
available for that day.
}
