% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o3}
\alias{cqcp_o3}
\title{Optional QC step o3}
\usage{
cqcp_o3(data, cutOff = 0.8)
}
\arguments{
\item{data}{data.table as returned from o2}

\item{cutOff}{percentage of values that must be present for each month before
all values of that month are flagged with FALSE, expressed in fraction: 0 <
cutOff < 1. Default is 0.8, i.e, 80 percent of data.}
}
\value{
data.table
}
\description{
Optional QC for temporal data availability. Flags all values in a month
as FALSE if less than 'cutOff' percent of valid values are available for that
month.
}
\examples{
o_3 <- o3(o_2)
}
