% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o4}
\alias{cqcp_o4}
\title{Optional QC step o4}
\usage{
cqcp_o4(data, time_constant)
}
\arguments{
\item{data}{data.table as returned from o1, o2, or o3 (column ta_int must be present)}

\item{time_constant}{time constant value for the sensor in seconds (must be 
the same for all p_id). Time constant/tau is defined as the time for a 
system's step response to ≈ 63.2 % of its final (asymptotic) value (from a 
step change) (https://en.wikipedia.org/wiki/Time_constant).}
}
\value{
data.table
}
\description{
Optional QC for data correction. Corrects all values for a known time constant 
(tau) of the sensor. This value must be the same for all stations/p_id and it is 
assumed that tau is constant, regardless of weather conditions. In the 
correction it is assumed that a step change in the values happens at each time 
step to the next.
}
\examples{
o_4 <- cqcp_o4(o_3, 1480.5)
}
