% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_output_statistics}
\alias{cqcp_output_statistics}
\title{Simple statistics on data availability after CrowdQC+}
\usage{
cqcp_output_statistics(data, print = TRUE, file = NULL)
}
\arguments{
\item{data}{data.table after CrowdQC+}

\item{print}{Set to TRUE to print information in the console. Default: TRUE}

\item{file}{Provide a file path to store the information in a file.}
}
\value{
data.table with output statistics
}
\description{
Calculate how many valid values remain after each QC step and print the
information.
}
