% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_padding}
\alias{cqcp_padding}
\title{Padding of data for application of CrowdQC+}
\usage{
cqcp_padding(
  data,
  resolution = "1 hour",
  rounding_method = "nearest",
  quiet = FALSE
)
}
\arguments{
\item{data}{data.table with at least columns 'p_id', 'time, 'ta'}

\item{resolution}{Temporal resolution as supported by lubridate. Default: '1 hour'}

\item{rounding_method}{Method to apply to round the time values to (cf. lubridate). 
Default is 'nearest', to assign the values to the nearest full time value, 
defined by 'resolution'. Other options are 'ceiling'/'ceil' and 'floor'.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table with regular time series for all stations.
}
\description{
CrowdQC+ (as CrowdQC) requires that all stations in the data.table have the 
same length and the same temporal resolution. Missing times are set to NaN.
This can be done using cqcp_padding. Currently, time stamps are rounded and
no interpolation is carried out to the data itself.
}
