% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_qcCWS}
\alias{cqcp_qcCWS}
\title{Complete quality control (QC) of CWS data}
\usage{
cqcp_qcCWS(
  data,
  m1_cutOff = 1,
  m2_low = 0.1,
  m2_high = 0.95,
  m2_lapse_rate = 0.006,
  m2_fun = qnorm,
  m3_cutOff = 0.2,
  m4_cutOff = 0.9,
  o1_fun = interpol,
  o2_cutOff = 0.8,
  o3_cutOff = 0.8,
  includeOptional = T,
  ...
)
}
\arguments{
\item{data}{input data in the format as the example data (netatmoBer)}

\item{m1_cutOff}{see cutOff in ?m1}

\item{m2_low}{see low in ?m2}

\item{m2_high}{see high in ?m2}

\item{m2_lapse_rate}{see lapse_rate in ?m2}

\item{m2_fun}{see fun in ?m2}

\item{m3_cutOff}{see cutOff in ?m3}

\item{m4_cutOff}{see cutOff in ?m4}

\item{o1_fun}{see fun in ?o1}

\item{o2_cutOff}{see cutOff in ?o2}

\item{o3_cutOff}{see cutOff in ?o3}

\item{includeOptional}{set to TRUE if QC steps o1 to o3 shall be
performed, default: TRUE}

\item{...}{additional parameters used in o1. For details see ?o1}
}
\value{
data.table
}
\description{
Performs all QC steps in consecutive order. All settings are according
to Napoly et al. (2018). This is the default function to carry out the
complete QC procedure. Each QC step takes the result of the previous
QC step as input. Thus, e.g., when applying QC step m2, the column
'm1' must be in the input data, for QC step m3 the column 'm2' must be
present, etc. Each individual step could be skipped by renaming the columns
in the input data. After each QC step a new column of type BYTE is
included in the output, containing TRUE or FALSE flags. Flags of the previous
levels are carried along, i.e., if a value failed in step m2, this FALSE is
kept throughout the remaining QC steps. In the end, only those data
values containing TRUE after the all QC steps are valid according to this
QC.
}
\examples{
data(netatmoBer)
y <- cqcp_qcCWS(netatmoBer)
}
