% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_Qnr}
\alias{cqcp_Qnr}
\title{Version of the Qn function respecting NaN values}
\usage{
cqcp_Qnr(x)
}
\arguments{
\item{x}{}
}
\value{
NaN for no valid values, otherwise robust scale estimator 'Qn' without NaN
}
\description{
Version of the Qn function respecting NaN values
}
