% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_add_dem_height}
\alias{cqcp_add_dem_height}
\title{Add DEM height to data.table.}
\usage{
cqcp_add_dem_height(
  data,
  file = NULL,
  raster = NULL,
  directory = NULL,
  outfile = NULL,
  overwrite = TRUE,
  crop = FALSE,
  na_vals = 0,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{data.table/frame with at least columns 'lon' and 'lat'}

\item{file}{Path to a DEM geotiff}

\item{raster}{RasterLayer object with DEM data}

\item{directory}{Directory path to store the downloaded SRTM data. If NULL,
downloaded data is stored in the current working directory.}

\item{outfile}{File path to save the created SRTM raster as geotiff}

\item{overwrite}{overwrite existing geotiff?}

\item{crop}{Crop SRTM raster/geotiff to data extent}

\item{na_vals}{Set NA values in DEM to this value to avoid missing value in 
cqcp_m2. Default: 0}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}

\item{...}{Additional parameters supported by geodata::elevation_3s}
}
\value{
data table with new column 'z' with DEM information
}
\description{
Adds Digital Elevation Model (DEM) information from a RasterLayer Object or 
geotiff to the data.table.
If no RasterLayer Object or file are given as parameters, SRTM data 
(Jarvis et al. 2008) is automatically downloaded and used to extract DEM 
information at the positions of the stations. Keep in mind that SRTM data is 
only available for regions between 60° N and 56° S. For regions outside that 
range, or if the user wants to use another DEM, that DEM can be used as input 
via the 'file' parameter (path to geotiff) or via the 'raster' parameter 
(RasterLayer Object).
}
\details{
A new column 'z' is added to the data.table to be used in QC level m2.

Reference: 
Jarvis A., Reuter, H. I., Nelson, A. and Guevara, E. (2008): Hole-filled 
seamless SRTM data V4, International Centre for Tropical Agriculture (CIAT), 
available from https://srtm.csi.cgiar.org
}
