% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_helper.R
\name{cqcp_check_input}
\alias{cqcp_check_input}
\title{Input check of data for CrowdQC+}
\usage{
cqcp_check_input(data, print = TRUE, file = NULL)
}
\arguments{
\item{data}{data.table to be checked.}

\item{print}{Set to TRUE to print information in the console. Default: TRUE}

\item{file}{Provide a file path to store the information in a file.}
}
\value{
logical, TRUE if data passed all checks.
}
\description{
Checks if a data.table has the required format for CrowdQC+ and gives hints
regarding the application of the package.
}
\details{
1. Check for column names.
2. Check for same temporal coverage.
3. Check for data regularity.
4. Geographical extent of data.
5. Number of stations.
}
\examples{
data(CWSBer)
cqcp_check_input(CWSBer)
}
