% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m1}
\alias{cqcp_m1}
\title{Main QC step m1}
\usage{
cqcp_m1(data, cutOff = 1, quiet = FALSE)
}
\arguments{
\item{data}{A data set (a data.table object) with required columns p_id, time, ta, lon, lat}

\item{cutOff}{How much stations are allowed to have the same coordinates,
default is 1. This means that if two p_ids share the same lon & lat values,
data at these stations are set to FALSE.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Flag values with FALSE that are missing or belong to a p_id having lon & lat
values which occur more often than the 'cutOff' value.
}
