% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m4}
\alias{cqcp_m4}
\title{Main QC step m4}
\usage{
cqcp_m4(data, cutOff = 0.9, complete = FALSE, duration = NULL, quiet = FALSE)
}
\arguments{
\item{data}{data.table object as returned by cqcp_m3}

\item{cutOff}{Value of correlation coefficient below which data are flagged
with FALSE, 0 < cutOff < 1. Default is 0.9.}

\item{complete}{Set to TRUE to use the complete data set for the filter 
(priority over 'duration').}

\item{duration}{A fixed duration to be used (cf. lubridate duration documentation).
This can be, e.g., '10 days'.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Flag values with FALSE if the correlation with the median of all stations is 
lower than cutOff.
Default is to apply m4 per month, but can be changed to use the complete data
set (e.g. for short data sets, 'complete = T'), or some other fixed duration 
(e.g. 'duration = "10 days"').
}
\examples{
data(cqcp_cws_data)
data <- cqcp_padding(cqcp_cws_data)
data_qc <- cqcp_m1(data)
data_qc <- cqcp_m2(data_qc)
data_qc <- cqcp_m3(data_qc)
data_qc <- cqcp_m4(data_qc)
}
