% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m5}
\alias{cqcp_m5}
\title{Main QC step m5}
\usage{
cqcp_m5(
  data,
  radius = 3000,
  n_station = 5,
  multiple_sd = 3,
  heightCorrection = T,
  lapse_rate = 0.0065
)
}
\arguments{
\item{data}{data.table as returned by m4}

\item{radius}{A radius in m around each station to check for neighbours.}

\item{n_station}{Minimum number of neighbouring stations with a valid value 
within radius.}

\item{multiple_sd}{Multiples of the spatial standard deviation within radius to 
be applied in the filter.}

\item{heightCorrection}{if set to true (default) and the column "z" exists in
the input data, the temperatures used in calculating the z-score are
corrected. The applied formula is ta_cor = ta + ((lapse_rate * (z - mz)) 
where mz is the spatial mean of z at each time step.}

\item{lapse_rate}{Lapse rate to use in 'heightCorrection'. Default is the 
environmental lapse rate of -0.0065 K/m. Set as a positive value: e.g.
lapse_rate = 0.01 to set a dry adiabatic lapse rate. For consistency, this
lapse rate should be the same as in ?m2.}
}
\value{
data.table
}
\description{
Spatial buddy check. Flags all values as FALSE if values are too different
than those from spatially neighbouring stations.
Check is based on the mean value plus/minus multiples of the standard deviation 
(default is three standard deviations) of all stations within a specified radius 
(default is 3000 m). A minimum number of stations within a radius have to be 
present for the calculations (default is five neighbours), thus this filter 
also flags isolated stations.
}
