% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_m5}
\alias{cqcp_m5}
\title{Main QC step m5}
\usage{
cqcp_m5(
  data,
  radius = 3000,
  n_station = 5,
  cutoff = 2,
  heightCorrection = TRUE,
  lapse_rate = 0.0065,
  check_elevation = TRUE,
  max_elev_diff = 100
)
}
\arguments{
\item{data}{data.table as returned by m4}

\item{radius}{A radius in meter around each station to check for neighbours.
Default: 3000}

\item{n_station}{Minimum number of neighbouring stations with a valid 
value within 'radius' Default: 5}

\item{cutoff}{Cutoff value for the outlier detection within 'radius' to be 
applied. Default: 2}

\item{heightCorrection}{If set to TRUE (default) and the column "z" exists in
the input data, the temperatures used in calculating the z-score are
corrected. The applied formula is ta_cor = ta + ((lapse_rate * (z - mz)) 
where mz is the spatial mean of z at each time step.}

\item{lapse_rate}{Lapse rate to use in 'heightCorrection'. Default is the 
environmental lapse rate of -0.0065 K/m. Set as a positive value: e.g.
lapse_rate = 0.01 to set a dry adiabatic lapse rate. For consistency, this
lapse rate should be the same as in ?cqcp_m2.}

\item{check_elevation}{Set to TRUE to check whether the elevation in column "z" 
of the neighbouring stations within 'radius' should be compared to the 
elevation of the station that is checked. If the absolute elevation difference 
between each buddy and the station itself is larger than 'max_elev_diff', this 
neighbour is left out of the calculations. Default: TRUE}

\item{max_elev_diff}{Maximum allowed elevation difference in meters between 
the station that is checked and each neighbour within 'radius'. Default: 100 m.}
}
\value{
data.table
}
\description{
Spatial buddy check. Flags all values as FALSE if values are too different
than those from spatially neighbouring stations (spatial outlier filter).
Check is based on the median value and the Qn estimator, as in cqcp_m2, of 
the buddies within a specified radius. The station that is checked is not
considered when calculating these values.
A minimum number of buddies within the radius has to be present for the 
calculations, thus this filter also flags isolated stations.
}
\examples{
data(netatmoBer)
m_1 <- cqcp_m1(netatmoBer)
m_2 <- cqcp_m2(m_1)
m_3 <- cqcp_m3(m_2)
m_4 <- cqcp_m4(m_3)
m_5 <- cqcp_m5(m_4)
}
