% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o1}
\alias{cqcp_o1}
\title{Optional QC step o1}
\usage{
cqcp_o1(data, fun = cqcp_interpol, quiet = FALSE, ...)
}
\arguments{
\item{data}{data.table as returned from m4}

\item{fun}{Function to use for interpolation, default is interpol}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}

\item{...}{Additional parameters for interpolation function}
}
\value{
data.table
}
\description{
In this step missing data is interpolated, default is to perform linear
interpolation on gaps of maximal length = 1.
A new column named "ta_int" is added to the data.table, containing the
filtered data at level m4 with the interpolated data.
}
\examples{
#default
data(CWSBer)
m_1 <- cqcp_m1(CWSBer)
m_2 <- cqcp_m2(m_1)
m_3 <- cqcp_m3(m_2)
m_4 <- cqcp_m4(m_3)
m_5 <- cqcp_m5(m_4)
o_1 <- cqcp_o1(m_5)
#interpolate gaps up to 5 hours
o_1 <- cqcp_o1(m_5, maxLength = 5)
}
