% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o2}
\alias{cqcp_o2}
\title{Optional QC step o2}
\usage{
cqcp_o2(data, cutOff = 0.8, quiet = FALSE)
}
\arguments{
\item{data}{data.table object as returned by cqcp_o1}

\item{cutOff}{Percentage of values that must be present for each day before
all values of that day are flagged with FALSE, expressed in fraction: 0 <
cutOff < 1. Default is 0.8, i.e, 80 percent of data.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Optional QC for temporal data availability. Flags all values in a
calendar day as FALSE if less than 'cutOff' percent of valid values are
available for that day.
}
\examples{
data(CWSBer)
m_1 <- cqcp_m1(CWSBer)
m_2 <- cqcp_m2(m_1)
m_3 <- cqcp_m3(m_2)
m_4 <- cqcp_m4(m_3)
m_5 <- cqcp_m5(m_4)
o_1 <- cqcp_o1(m_5)
o_2 <- cqcp_o2(o_1)
}
