% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o3}
\alias{cqcp_o3}
\title{Optional QC step o3}
\usage{
cqcp_o3(data, cutOff = 0.8, complete = FALSE, duration = NULL, quiet = FALSE)
}
\arguments{
\item{data}{data.table object as returned by cqcp_o2}

\item{cutOff}{Percentage of values that must be present for each duration before
all values of that duration are flagged with FALSE, expressed in fraction: 0 <
cutOff < 1. Default is 0.8, i.e, 80 percent of data.}

\item{complete}{Set to TRUE to use the complete data set for the filter 
(priority over 'duration').}

\item{duration}{A fixed duration to be used (cf. lubridate duration documentation).
This can be, e.g., '10 days'.}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Optional QC for temporal data availability. Flags all values in a specified 
duration (if nothing specified on a monthly basis) as FALSE if less than 
'cutOff' percent of valid values are available for that duration.
}
