% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_o4}
\alias{cqcp_o4}
\title{Optional QC step o4}
\usage{
cqcp_o4(data, time_constant, quiet = FALSE)
}
\arguments{
\item{data}{data.table in CrowdQC+ format (columns "time" and "ta" must be present)}

\item{time_constant}{Time constant value for the sensor in seconds (must be 
the same for all p_id). Time constant/tau is defined as the time for a 
system's step response to ≈ 63.2 % of its final (asymptotic) value (from a 
step change) (https://en.wikipedia.org/wiki/Time_constant).}

\item{quiet}{Suppress messages by CrowdQC+. Default: FALSE}
}
\value{
data.table
}
\description{
Optional QC for data correction. Corrects all values for a known time constant 
(tau) of the sensor. This value must be the same for all stations/p_id and it is 
assumed that tau is constant, regardless of weather conditions. In the 
correction it is assumed that a step change in the values happens at each time 
step to the next.
In the correction the original data "ta" is used instead of the interpolated
values "ta_int". Hence, this function can be applied at/after any QC level. 
Diverging from all other QC levels, no additional flag variable with TRUE/FALSE 
is added to the data.table during cqcp_o4. Corrected data can thus be 
selected at any QC level. The corrected data are written in a new column 'ta_corr'. 
For Netatmo CWS (https://www.netatmo.com/en-us/weather), Coney et al. 2022 
(https://doi.org/10.1002/met.2075) determined a mean time constant of the 
air-temperature sensor of tau = 12.7 minutes (762 seconds).
Büchau 2018 (MSc thesis, https://bis-erdsystem.de/fileadmin/user_upload/bise/Texte/MSC_Met_61-20180313.pdf)
determined a mean time constant of the Netatmo air-temperature sensor (including 
silver aluminium shell) of tau = 24.675 minutes (1480.5 seconds).
}
