% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqcp_filter.R
\name{cqcp_qcCWS}
\alias{cqcp_qcCWS}
\title{Complete quality control (QC) of CWS data}
\usage{
cqcp_qcCWS(
  data,
  m1_cutOff = 1,
  m2_low = 0.01,
  m2_high = 0.95,
  m2_heightCorrection = TRUE,
  m2_debug = FALSE,
  m2_lapse_rate = 0.0065,
  m2_t_distribution = FALSE,
  m3_cutOff = 0.2,
  m4_cutOff = 0.9,
  m5_radius = 3000,
  m5_n_buddies = 5,
  m5_alpha = 0.1,
  m5_heightCorrection = TRUE,
  m5_lapse_rate = 0.0065,
  m5_check_elevation = TRUE,
  m5_max_elev_diff = 100,
  m5_keep_isolated = FALSE,
  o1_fun = cqcp_interpol,
  o2_cutOff = 0.8,
  o3_cutOff = 0.8,
  o4_time_constant = NULL,
  includeOptional = TRUE,
  complete = FALSE,
  duration = NULL,
  ...
)
}
\arguments{
\item{data}{Input data in the format as the example data (CWSBer)}

\item{m1_cutOff}{see cutOff in ?cqcp_m1}

\item{m2_low}{see low in ?cqcp_m2}

\item{m2_high}{see high in ?cqcp_m2}

\item{m2_heightCorrection}{see heightCorrection in ?cqcp_m2}

\item{m2_debug}{see debug in ?cqcp_m2}

\item{m2_lapse_rate}{see lapse_rate in ?cqcp_m2}

\item{m2_t_distribution}{see t_distribution in ?cqcp_m2}

\item{m3_cutOff}{see cutOff in ?cqcp_m3}

\item{m4_cutOff}{see cutOff in ?cqcp_m4}

\item{m5_radius}{see radius in ?cqcp_m5}

\item{m5_n_buddies}{see n_buddies in ?cqcp_m5}

\item{m5_alpha}{see alpha in ?cqcp_m5}

\item{m5_heightCorrection}{see heightCorrection in ?cqcp_m5}

\item{m5_lapse_rate}{see lapse_rate in ?cqcp_m5}

\item{m5_check_elevation}{see check_elevation in ?cqcp_m5}

\item{m5_max_elev_diff}{see max_elev_diff in ?cqcp_m5}

\item{m5_keep_isolated}{see keep_isolated in ?cqcp_m5}

\item{o1_fun}{see fun in ?cqcp_o1}

\item{o2_cutOff}{see cutOff in ?cqcp_o2}

\item{o3_cutOff}{see cutOff in ?cqcp_o3}

\item{o4_time_constant}{see time_constant in ?cqcp_o4}

\item{includeOptional}{set to TRUE if QC steps cqcp_o1 to cqcp_o3 shall be
performed, default: TRUE}

\item{complete}{see complete in ?cqcp_m3, ?cqcp_m4, or ?cqcp_o3}

\item{duration}{see duration in ?cqcp_m3, ?cqcp_m4, or ?cqcp_o3}

\item{...}{Additional parameters used in cqcp_o1. For details see ?cqcp_o1}
}
\value{
data.table
}
\description{
Performs all QC steps in consecutive order. All settings are according
to Napoly et al. (2018). This is the default function to carry out the
complete QC procedure. Each QC step takes the result of the previous
QC step as input. Thus, e.g., when applying QC step m2, the column
'm1' must be in the input data, for QC step m3 the column 'm2' must be
present, etc. Each individual step could be skipped by renaming the columns
in the input data. After each QC step a new column of type BYTE is
included in the output, containing TRUE or FALSE flags. Flags of the previous
levels are carried along, i.e., if a value failed in step m2, this FALSE is
kept throughout the remaining QC steps. In the end, only those data
values containing TRUE after the all QC steps are valid according to this
QC.
'complete' currently does not work with the example data set (CWSBer).
}
\examples{
data(CWSBer)
y <- cqcp_qcCWS(CWSBer)
}
